better docs route
